xilinx-zymqmp: Add ostree support
The specific variables required to properly boot the device have been
added.  It is possible to network install directly to an existing mmc
card, so long as there is free space on the device that is not
partitioned.  If the fat32 partition takes up the entire mmc card it
is possible to live repartition it, but that is a manual process,
similar to what you would use to resize a live partition on a typical
linux distribution.

[ Issue: LINUXEXEC-3408 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>

[ Issue: LINUXEXEC-4069 ]

Get this patch from BSP xilinx-zymqmp, branch is WRLINUX_10_18_LTS.

Signed-off-by: Meng Li <Meng.Li@windriver.com>