xilinx-zynqmp: zynqmp-secure-boot: adjust for LTS21
Issue: LINUXEXEC-17321

There are modifications as below:
1. In zynqmp-secure-boot.inc, add UBOOT_ELF/UBOOT_ELF_SUFFIX so that
u-boot.elf will be installed to tmp-glibc/deploy/images/xilinx-zynqmp.
2. In u-boot-xlnx.inc, redefine ${B} and ${UBOOT_INITIAL_ENV} to fix
incompatibility since u-boot.inc in meta layer is for 2021.1 but we
are using 2019.1 xilinx u-boot.
3. In linux-yocto_5.10.bbappend, change all "xx@xx" to "xx-xx" according
to the commit 6047be9f8f ("kernel-fitimage: Don't use unit addresses on
FIT") in oe-core layer. This is because uboot-mkimage is provided by
u-boot-tools_2021.01 which contains the upstream u-boot commit
3f04db891a35 ("image: Check for unit addresses in FITs").
4. In linux-yocto_5.10.bbappend, fitimage_emit_section_config will be
passed with a new argument set because of the oe-core commit 19fa415c87
("kernel-fitimage: adding support for Initramfs bundle and u-boot
 script").

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>